requirements.txt
reportlab
num2words
requests
pyodbc
fastapi
uvicorn
jinja2
python-dotenv
